+++ lib/src/executable/start_command.dart
@@ -1,3 +1,4 @@
+import 'package:args/command_runner.dart';
 import 'package:virtual_device/src/executable/executable_utils.dart';
 import 'package:virtual_device/src/executable/virtual_device_command.dart';
 
@@ -8,19 +9,24 @@ class StartCommand extends VirtualDeviceCommand {
   @override
   final description = 'Boot an existing virtual device';
 
-  StartCommand() {
-    argParser.addOption(
-      'name',
-      help: 'The device name.',
-    );
-  }
+  @override
+  final usage = 'Specify device names to start in a space-separated list';
 
   @override
   Future<void> run() async {
-    final device = await deviceFromName(
-      isIOS: isIos,
-      name: argResults['name'],
-    );
-    return await device.start();
+    if (argResults.rest?.isEmpty ?? true) {
+      throw UsageException(
+        'The name of the device(s) is required',
+        'Ex. virtual_device start "iPad Air 2:12.1:1"',
+      );
+    }
+
+    for (final name in argResults.rest) {
+      final device = await deviceFromName(
+        isIOS: isIos,
+        name: name,
+      );
+      await device.start();
+    }
   }
 }